refactor controller code to service class
